DEV: Move logic into a service
